Создает "заготовку" моков для приложения. Переносит '__test-data__'  в папку 'src'.
