+++ package.json
@@ -37,7 +37,7 @@
   },
   "jest": {
     "setupFilesAfterEnv": [
-      "<rootDir>/src/components/__test-data__/test-setup.js"
+      "<rootDir>/src/__test-data__/test-setup.js"
     ]
   }
 }
