+++ src/components/catalog-genres-item/catalog-genres-item.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 import {shallow} from "enzyme";
 import {CatalogGenresItem} from "./catalog-genres-item.jsx";
-import {genreName} from "../__test-data__/test-mocks.js";
+import {genreName} from "../../__test-data__/test-mocks.js";
 
 
 const props = {
